Uses callback to set github auth
The async callback previously used to get the ssm parameters caused
a race condition on the calls to the github api. If the ssm calls
returned quickly enough (before the github api calls executed), it
worked. However, if the github api calls executed first, then the
calls would be unauthenticated and result in an error.